SAN-219 Update HandleHealthCheckFinanceSystem to use new IssuerGatewayHealthcheckFinanceSystem - CheckScheduledMaintenance
