+++ .dockerignore
@@ -7,6 +7,7 @@
 !e5
 !handlers
 !interceptors
+!issuer_gateway
 !middleware
 !mocks
 !models
@@ -18,4 +19,3 @@
 !go.mod
 !go.sum
 !main.go
-!issuer_gateway
