+++ issuer_gateway/api/healthcheck_finance.go
@@ -1,0 +1,9 @@
+package api
+
+import (
+	"time"
+)
+
+type HealthcheckFinanceSystem interface {
+	CheckScheduledMaintenance() (time.Time, bool, bool)
+}
